cmd: move simple session factory to cmd/utils.go
