+++ pkg/cmd/balances.go
@@ -6,10 +6,6 @@ import (
 
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-
-	"github.com/c9s/bbgo/pkg/exchange/ftx"
-	"github.com/c9s/bbgo/pkg/types"
 )
 
 //godotenv -f .env.local go run ./cmd/bbgo balances --session=ftx
@@ -40,17 +36,4 @@ func init() {
 	balancesCmd.Flags().String("session", "", "the exchange session name for sync")
 
 	RootCmd.AddCommand(balancesCmd)
-}
-
-func newExchange(session string) (types.Exchange, error) {
-	switch session {
-	case "ftx":
-		return ftx.NewExchange(
-			viper.GetString("ftx-api-key"),
-			viper.GetString("ftx-api-secret"),
-			viper.GetString("ftx-subaccount-name"),
-		), nil
-
-	}
-	return nil, fmt.Errorf("unsupported session %s", session)
 }
